Expose optional tuning params
