+++ packages/color/docs/markdown/color.tuningparameters.analogousrange.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [analogousRange](./color.tuningparameters.analogousrange.md)
+
+## TuningParameters.analogousRange property
+
+<b>Signature:</b>
+
+```typescript
+analogousRange: number;
+```
