+++ packages/color/docs/markdown/color.tuningparameters.backgroundhueshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [backgroundHueShift](./color.tuningparameters.backgroundhueshift.md)
+
+## TuningParameters.backgroundHueShift property
+
+<b>Signature:</b>
+
+```typescript
+backgroundHueShift: number;
+```
