+++ packages/color/docs/markdown/color.tuningparameters.backgroundlevel.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [backgroundLevel](./color.tuningparameters.backgroundlevel.md)
+
+## TuningParameters.backgroundLevel property
+
+<b>Signature:</b>
+
+```typescript
+backgroundLevel: number;
+```
