+++ packages/color/docs/markdown/color.tuningparameters.complementaryrange.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [complementaryRange](./color.tuningparameters.complementaryrange.md)
+
+## TuningParameters.complementaryRange property
+
+<b>Signature:</b>
+
+```typescript
+complementaryRange: number;
+```
