+++ packages/color/docs/markdown/color.tuningparameters.darkbackgroundlightnessshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [darkBackgroundLightnessShift](./color.tuningparameters.darkbackgroundlightnessshift.md)
+
+## TuningParameters.darkBackgroundLightnessShift property
+
+<b>Signature:</b>
+
+```typescript
+darkBackgroundLightnessShift: number;
+```
