+++ packages/color/docs/markdown/color.tuningparameters.darkestgrey.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [darkestGrey](./color.tuningparameters.darkestgrey.md)
+
+## TuningParameters.darkestGrey property
+
+<b>Signature:</b>
+
+```typescript
+darkestGrey: number;
+```
