+++ packages/color/docs/markdown/color.tuningparameters.darkmaxlightnessoffet.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [darkMaxLightnessOffet](./color.tuningparameters.darkmaxlightnessoffet.md)
+
+## TuningParameters.darkMaxLightnessOffet property
+
+<b>Signature:</b>
+
+```typescript
+darkMaxLightnessOffet: number;
+```
