+++ packages/color/docs/markdown/color.tuningparameters.darktextlightness.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [darkTextLightness](./color.tuningparameters.darktextlightness.md)
+
+## TuningParameters.darkTextLightness property
+
+<b>Signature:</b>
+
+```typescript
+darkTextLightness: number;
+```
