+++ packages/color/docs/markdown/color.tuningparameters.lightbackgroundlightnessshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [lightBackgroundLightnessShift](./color.tuningparameters.lightbackgroundlightnessshift.md)
+
+## TuningParameters.lightBackgroundLightnessShift property
+
+<b>Signature:</b>
+
+```typescript
+lightBackgroundLightnessShift: number;
+```
