+++ packages/color/docs/markdown/color.tuningparameters.lightestgrey.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [lightestGrey](./color.tuningparameters.lightestgrey.md)
+
+## TuningParameters.lightestGrey property
+
+<b>Signature:</b>
+
+```typescript
+lightestGrey: number;
+```
