+++ packages/color/docs/markdown/color.tuningparameters.lightmaxlightnessoffset.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [lightMaxLightnessOffset](./color.tuningparameters.lightmaxlightnessoffset.md)
+
+## TuningParameters.lightMaxLightnessOffset property
+
+<b>Signature:</b>
+
+```typescript
+lightMaxLightnessOffset: number;
+```
