+++ packages/color/docs/markdown/color.tuningparameters.lighttextlightness.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [lightTextLightness](./color.tuningparameters.lighttextlightness.md)
+
+## TuningParameters.lightTextLightness property
+
+<b>Signature:</b>
+
+```typescript
+lightTextLightness: number;
+```
