+++ packages/color/docs/markdown/color.tuningparameters.lowcontrastbackgroundshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [lowContrastBackgroundShift](./color.tuningparameters.lowcontrastbackgroundshift.md)
+
+## TuningParameters.lowContrastBackgroundShift property
+
+<b>Signature:</b>
+
+```typescript
+lowContrastBackgroundShift: number;
+```
