+++ packages/color/docs/markdown/color.tuningparameters.maxbackgroundchroma.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [maxBackgroundChroma](./color.tuningparameters.maxbackgroundchroma.md)
+
+## TuningParameters.maxBackgroundChroma property
+
+<b>Signature:</b>
+
+```typescript
+maxBackgroundChroma: number;
+```
