+++ packages/color/docs/markdown/color.tuningparameters.md
@@ -1,0 +1,43 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md)
+
+## TuningParameters interface
+
+Detailed tuning for the scale generation algorithms.
+
+<b>Signature:</b>
+
+```typescript
+export interface TuningParameters 
+```
+
+## Properties
+
+|  Property | Modifiers | Type | Description |
+|  --- | --- | --- | --- |
+|  [analogousRange](./color.tuningparameters.analogousrange.md) |  | number |  |
+|  [backgroundHueShift](./color.tuningparameters.backgroundhueshift.md) |  | number |  |
+|  [backgroundLevel](./color.tuningparameters.backgroundlevel.md) |  | number |  |
+|  [complementaryRange](./color.tuningparameters.complementaryrange.md) |  | number |  |
+|  [darkBackgroundLightnessShift](./color.tuningparameters.darkbackgroundlightnessshift.md) |  | number |  |
+|  [darkestGrey](./color.tuningparameters.darkestgrey.md) |  | number |  |
+|  [darkMaxLightnessOffet](./color.tuningparameters.darkmaxlightnessoffet.md) |  | number |  |
+|  [darkTextLightness](./color.tuningparameters.darktextlightness.md) |  | number |  |
+|  [lightBackgroundLightnessShift](./color.tuningparameters.lightbackgroundlightnessshift.md) |  | number |  |
+|  [lightestGrey](./color.tuningparameters.lightestgrey.md) |  | number |  |
+|  [lightMaxLightnessOffset](./color.tuningparameters.lightmaxlightnessoffset.md) |  | number |  |
+|  [lightTextLightness](./color.tuningparameters.lighttextlightness.md) |  | number |  |
+|  [lowContrastBackgroundShift](./color.tuningparameters.lowcontrastbackgroundshift.md) |  | number |  |
+|  [maxBackgroundChroma](./color.tuningparameters.maxbackgroundchroma.md) |  | number |  |
+|  [minNominalLightness](./color.tuningparameters.minnominallightness.md) |  | number |  |
+|  [minNominalSaturation](./color.tuningparameters.minnominalsaturation.md) |  | number |  |
+|  [nominalBoldLightnessShift](./color.tuningparameters.nominalboldlightnessshift.md) |  | number |  |
+|  [nominalBoldSaturationShift](./color.tuningparameters.nominalboldsaturationshift.md) |  | number |  |
+|  [nominalHueStep](./color.tuningparameters.nominalhuestep.md) |  | number |  |
+|  [nominalMutedLightnessShift](./color.tuningparameters.nominalmutedlightnessshift.md) |  | number |  |
+|  [nominalMutedSaturationShift](./color.tuningparameters.nominalmutedsaturationshift.md) |  | number |  |
+|  [offsetBackgroundLightnessShift](./color.tuningparameters.offsetbackgroundlightnessshift.md) |  | number |  |
+|  [polynomialExponent](./color.tuningparameters.polynomialexponent.md) |  | number |  |
+|  [reservedDataColors](./color.tuningparameters.reserveddatacolors.md) |  | number |  |
+
