+++ packages/color/docs/markdown/color.tuningparameters.minnominallightness.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [minNominalLightness](./color.tuningparameters.minnominallightness.md)
+
+## TuningParameters.minNominalLightness property
+
+<b>Signature:</b>
+
+```typescript
+minNominalLightness: number;
+```
