+++ packages/color/docs/markdown/color.tuningparameters.minnominalsaturation.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [minNominalSaturation](./color.tuningparameters.minnominalsaturation.md)
+
+## TuningParameters.minNominalSaturation property
+
+<b>Signature:</b>
+
+```typescript
+minNominalSaturation: number;
+```
