+++ packages/color/docs/markdown/color.tuningparameters.nominalboldlightnessshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [nominalBoldLightnessShift](./color.tuningparameters.nominalboldlightnessshift.md)
+
+## TuningParameters.nominalBoldLightnessShift property
+
+<b>Signature:</b>
+
+```typescript
+nominalBoldLightnessShift: number;
+```
