+++ packages/color/docs/markdown/color.tuningparameters.nominalboldsaturationshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [nominalBoldSaturationShift](./color.tuningparameters.nominalboldsaturationshift.md)
+
+## TuningParameters.nominalBoldSaturationShift property
+
+<b>Signature:</b>
+
+```typescript
+nominalBoldSaturationShift: number;
+```
