+++ packages/color/docs/markdown/color.tuningparameters.nominalhuestep.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [nominalHueStep](./color.tuningparameters.nominalhuestep.md)
+
+## TuningParameters.nominalHueStep property
+
+<b>Signature:</b>
+
+```typescript
+nominalHueStep: number;
+```
