+++ packages/color/docs/markdown/color.tuningparameters.nominalmutedlightnessshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [nominalMutedLightnessShift](./color.tuningparameters.nominalmutedlightnessshift.md)
+
+## TuningParameters.nominalMutedLightnessShift property
+
+<b>Signature:</b>
+
+```typescript
+nominalMutedLightnessShift: number;
+```
