+++ packages/color/docs/markdown/color.tuningparameters.nominalmutedsaturationshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [nominalMutedSaturationShift](./color.tuningparameters.nominalmutedsaturationshift.md)
+
+## TuningParameters.nominalMutedSaturationShift property
+
+<b>Signature:</b>
+
+```typescript
+nominalMutedSaturationShift: number;
+```
