+++ packages/color/docs/markdown/color.tuningparameters.offsetbackgroundlightnessshift.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [offsetBackgroundLightnessShift](./color.tuningparameters.offsetbackgroundlightnessshift.md)
+
+## TuningParameters.offsetBackgroundLightnessShift property
+
+<b>Signature:</b>
+
+```typescript
+offsetBackgroundLightnessShift: number;
+```
