+++ packages/color/docs/markdown/color.tuningparameters.polynomialexponent.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [polynomialExponent](./color.tuningparameters.polynomialexponent.md)
+
+## TuningParameters.polynomialExponent property
+
+<b>Signature:</b>
+
+```typescript
+polynomialExponent: number;
+```
