+++ packages/color/docs/markdown/color.tuningparameters.reserveddatacolors.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/color](./color.md) &gt; [TuningParameters](./color.tuningparameters.md) &gt; [reservedDataColors](./color.tuningparameters.reserveddatacolors.md)
+
+## TuningParameters.reservedDataColors property
+
+<b>Signature:</b>
+
+```typescript
+reservedDataColors: number;
+```
