+++ packages/core/docs/markdown/core.themespec.md
@@ -20,4 +20,5 @@ export interface ThemeSpec
 |  [light?](./core.themespec.light.md) |  | Scheme | <i>(Optional)</i> |
 |  [name?](./core.themespec.name.md) |  | string | <i>(Optional)</i> |
 |  [params?](./core.themespec.params.md) |  | SchemeParams | <i>(Optional)</i> |
+|  [tuning?](./core.themespec.tuning.md) |  | TuningParameters | <i>(Optional)</i> |
