+++ packages/core/docs/markdown/core.themespec.tuning.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thematic/core](./core.md) &gt; [ThemeSpec](./core.themespec.md) &gt; [tuning](./core.themespec.tuning.md)
+
+## ThemeSpec.tuning property
+
+<b>Signature:</b>
+
+```typescript
+tuning?: TuningParameters;
+```
